# api/openapi.yaml
openapi: 3.0.3
info:
  title: Deployment Schedule API
  description: API for managing deployment schedules
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
paths:
...
  /groups:
    get:
      summary: List groups for an owner
      operationId: listGroups
      tags:
        - groups
      security:
        - bearerAuth: []
      parameters:
        - name: owner
          in: query
          required: true
          description: Filter groups by owner
          schema:
            type: string
      responses:
        '200':
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a new group
      operationId: createGroup
      tags:
        - groups
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Group name already exists for this owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups/{id}:
    get:
      summary: Get a group by ID
      operationId: getGroup
      tags:
        - groups
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Group details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update a group
      operationId: updateGroup
      tags:
        - groups
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupRequest'
      responses:
        '200':
          description: Group updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Group name already exists for this owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a group
      operationId: deleteGroup
      tags:
        - groups
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Group deleted
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups/{id}/favorite:
    post:
      summary: Mark a group as favorite
      operationId: favoriteGroup
      tags:
        - groups
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Group favorited successfully
        '401':
          description: Unauthorized - authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Remove favorite status from a group
      operationId: unfavoriteGroup
      tags:
        - groups
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Group unfavorited successfully
        '401':
          description: Unauthorized - authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /schedules/{id}/groups:
    post:

      summary: Assign schedule to groups

      operationId: assignScheduleToGroups

      tags:

        - schedules

      security:

        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Schedule ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignScheduleRequest'
      responses:
        '200':
          description: Schedule assigned to groups
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Schedule or group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: Get groups for a schedule
      operationId: getGroupsForSchedule
      tags:
        - schedules
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Schedule ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '404':
          description: Schedule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /schedules/{id}/groups/{groupId}:
    delete:

      summary: Unassign schedule from group

      operationId: unassignScheduleFromGroup

      tags:

        - schedules

      security:

        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Schedule ID
          schema:
            type: string
            format: uuid
        - name: groupId
          in: path
          required: true
          description: Group ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Schedule unassigned from group
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups/{id}/schedules:
    get:

      summary: Get schedules in a group

      operationId: getSchedulesInGroup

      tags:

        - groups

      security:

        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of schedule IDs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  format: uuid
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Bulk assign schedules to group
      operationId: bulkAssignSchedules
      tags:
        - groups
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkAssignRequest'
      responses:
        '200':
          description: Schedules assigned to group
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group or schedule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from Google OAuth login
  schemas:
...
    Group:
      type: object
      required:
        - id
        - name
        - visibility
        - owner
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique group identifier
        name:
          type: string
          maxLength: 100
          description: Name of the group
        description:
          type: string
          maxLength: 500
          description: Optional description of the group
        visibility:
          type: string
          enum: [public, private]
          description: Visibility of the group (public visible to all, private only to owner)
        owner:
          type: string
          description: Owner of the group (immutable after creation)
        createdAt:
          type: string
          format: date-time
          description: When the group was created
        updatedAt:
          type: string
          format: date-time
          description: When the group was last updated
        isFavorite:
          type: boolean
          description: Whether this group is favorited by the authenticated user
    CreateGroupRequest:
      type: object
      required:
        - name
        - owner
      properties:
        name:
          type: string
          maxLength: 100
          description: Name of the group
        description:
          type: string
          maxLength: 500
          description: Optional description of the group
        visibility:
          type: string
          enum: [public, private]
          default: private
          description: Visibility of the group (defaults to private)
        owner:
          type: string
          description: Owner of the group
    UpdateGroupRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 100
          description: Name of the group
        description:
          type: string
          maxLength: 500
          description: Optional description of the group
        visibility:
          type: string
          enum: [public, private]
          description: Visibility of the group
    AssignScheduleRequest:
      type: object
      required:
        - groupIds
      properties:
        groupIds:
          type: array
          items:
            type: string
            format: uuid
          description: Array of group IDs to assign the schedule to
        assignedBy:
          type: string
          description: User who performed the assignment
    BulkAssignRequest:
      type: object
      required:
        - scheduleIds
      properties:
        scheduleIds:
          type: array
          items:
            type: string
            format: uuid
          description: Array of schedule IDs to assign to the group
        assignedBy:
          type: string
          description: User who performed the assignment
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Error message
        details:
          type: string
          description: Additional error details